Fix: Collapsable Edit Menu Sections
Test Plan: Open the Welcome slide show. Click Edit. Open and close the
right hand sections such as "Align", "Text", "Lines", "Chords",
"Special", "CSS". Ensure that opening the first time set the value to
the default (for align this changes the text align from center to left
alignment - this was the previous behavior). Ensure that collapsing the
menu keeps the value it was overwritten to. Ensure that reopening the menu
keeps the value to was it was overwritten to. Ensure that clicking reset
converts the value back to the original value before the menu was opened
(note this may be different than the default for example with Align,
this was the previous behavior).

Note that "CSS" is missing from the language dictionary and thus
collapsing CSS exposes a race condition bug in that CSS will now show
"[edit.CSS]" instead of "CSS" upon collapse.