Adding transpose functionality to text edit (not working fully) (#1551)
* Fix makesnap (#1530)

The initiative to organize the scripts broke the snap building. Snaps require a very specific structure and the `snap` directory _must_ be in the root of the repository. WIthout this, it fundamentally breaks.

Additionally, the removal of the executable bit from the scripts caused the entire build to fail as the scripts _must_ be executable in order for the build to succeed and the installed snap to function at all.

I have revised the `makesnap` script to accommodate the changes and move the required files to a `snap` directory during build so it may build properly.

* feat: add arrow_up icon, ensure transpose UI uses consistent custom icons, and maintain code style in chord transposition logic

---------

Co-authored-by: Erich Eickmeyer <erich@ericheickmeyer.com>